crewai-service/requirements.txt
crewai>=0.108.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.8.0
